Move inline renderers to nested namespace
